Move to vendored_sdks
