+++ src/interactive/azext_interactive/azclishell/argfinder.py
@@ -7,7 +7,7 @@ import argparse
 
 from argcomplete import CompletionFinder
 
-from .compat import USING_PYTHON2, ensure_bytes
+from ..vendored_sdks.argcomplete.compat import USING_PYTHON2, ensure_bytes
 
 
 class ArgsFinder(CompletionFinder):  # pylint: disable=too-few-public-methods
