+++ src/interactive/azext_interactive/vendored_sdks/argcomplete/compat.py
